fix(hotfix-04): UE5 5.7 빌드 통과 (EClassId None + include path + Tests 캐스트) (#5)
UE5 에디터에서 'IdleProject could not be compiled' 에러로 PR #4 빌드 실패 → hotfix 3건:

1. CharacterSystem/StatFormulas.h: EClassId enum 에 None=0 추가
   - UHT (UnrealHeaderTool) -WarningsAsErrors 가 enum default(0) 값 부재를 fatal error 로 처리
   - UMETA(Hidden) 으로 BP 노출 차단

2. IdleProject.Build.cs: PublicIncludePaths.Add(ModuleDirectory)
   - 도메인 디렉터리 (CharacterSystem/, GameCore/, UI/ 등) 가 표준 Public/Private 구조가 아니므로
     모듈 루트 경로를 명시적으로 include base 로 추가하여 '#include "CharacterSystem/..."' 패턴 작동

3. Tests/StatFormulasTests.cpp, LevelFormulasTests.cpp: 명시적 캐스트
   - TestEqual 의 오버로드 모호 해결 — float 인자에 0.0f, int64 인자에 static_cast<int64> 적용

검증: Build.bat IdleProjectEditor Win64 Development → 'Result: Succeeded' (exit 0)
사용자 환경: VS Build Tools 17.14.33 + MSVC 14.44 + Win SDK 10.0.26100 + .NET Framework 4.8 SDK

+++ client/Source/IdleProject/Tests/LevelFormulasTests.cpp
@@ -10,17 +10,17 @@ IMPLEMENT_SIMPLE_AUTOMATION_TEST(
 
 bool FLevelFormulasExpTest::RunTest(const FString& Parameters)
 {
-	TestEqual(TEXT("LEVEL_CAP"), FLevelFormulas::LEVEL_CAP, 200);
-	TestEqual(TEXT("ExpToNext(1)"), FLevelFormulas::ExpToNext(1), 150);
-	TestEqual(TEXT("ExpToNext(10)"), FLevelFormulas::ExpToNext(10), 3506);
-	TestEqual(TEXT("ExpToNext(100)"), FLevelFormulas::ExpToNext(100), 135594);
-	TestEqual(TEXT("ExpToNext(200) cap sentinel"), FLevelFormulas::ExpToNext(200), 0);
+	TestEqual(TEXT("LEVEL_CAP"), FLevelFormulas::LEVEL_CAP, static_cast<int32>(200));
+	TestEqual(TEXT("ExpToNext(1)"), FLevelFormulas::ExpToNext(1), static_cast<int64>(150));
+	TestEqual(TEXT("ExpToNext(10)"), FLevelFormulas::ExpToNext(10), static_cast<int64>(3506));
+	TestEqual(TEXT("ExpToNext(100)"), FLevelFormulas::ExpToNext(100), static_cast<int64>(135594));
+	TestEqual(TEXT("ExpToNext(200) cap sentinel"), FLevelFormulas::ExpToNext(200), static_cast<int64>(0));
 
 	const int64 ExpectedLevel4Total =
 		FLevelFormulas::ExpToNext(1) +
 		FLevelFormulas::ExpToNext(2) +
 		FLevelFormulas::ExpToNext(3);
-	TestEqual(TEXT("CumulativeExp(1)"), FLevelFormulas::CumulativeExp(1), 0);
+	TestEqual(TEXT("CumulativeExp(1)"), FLevelFormulas::CumulativeExp(1), static_cast<int64>(0));
 	TestEqual(TEXT("CumulativeExp(4)"), FLevelFormulas::CumulativeExp(4), ExpectedLevel4Total);
 
 	return true;
